scraping/requirements.txt
amqp==5.1.1
appdirs==1.4.4
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
bs4==0.0.1
celery==5.3.4
celery-redbeat==2.1.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cssselect==1.2.0
fake-useragent==1.2.1
Flask==2.3.3
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
joblib==1.3.2
kombu==5.3.2
lxml==4.9.3
numpy==1.26.0
outcome==1.2.0
packaging==23.1
parse==1.19.1
prompt-toolkit==3.0.39
pyee==8.2.2
pyperclip==1.8.2
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
requests-html==0.10.0
scikit-learn==1.3.1
scipy==1.11.3
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
undetected-chromedriver==3.5.3
urllib3==1.26.16
vine==5.0.0
w3lib==2.1.2
wcwidth==0.2.8
webdriver-manager==4.0.0
websockets==10.4
wsproto==1.2.0
zipp==3.17.0